server/pyproject.toml

[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "server"
dynamic = ["version"]
requires-python = ">=3.10"

dependencies = [
  "torch>=1.13.1",
  "transformers>=4.43.0,<=4.45.0",
  "peft>=0.10.0",
  "loguru",
  "trl>=0.9.3,<=0.9.6",
  "bitsandbytes",
  "pyyaml",
  "fastapi",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "pre-commit",
  "pydata-sphinx-theme>=0.12",
  "pytest",
  "ruff",
  "sphinx-autobuild",
  "sphinx-copybutton",
  "sphinx-design",
  "tox",
  "tox-direct",
]

[tool.setuptools_scm]

[tool.pytest.ini_options]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
addopts = """
    --tb=native -vv --doctest-modules --doctest-glob="*.rst"
    """
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "tests"

# tox must currently be configured via an embedded ini string
# See: https://github.com/tox-dev/tox/issues/999
[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist=True

[testenv:{pre-commit,tests,docs}]
# Don't create a virtualenv for the command, requires tox-direct plugin
direct = True
passenv = *
allowlist_externals =
    pre-commit
    pytest
    sphinx-build
    sphinx-autobuild
commands =
    pre-commit: pre-commit run --all-files --show-diff-on-failure {posargs}
    tests: pytest {posargs}
    docs: sphinx-{posargs:build -EW --keep-going} -T docs build/html
"""

[tool.ruff]
src = ["src", "tests"]
line-length = 88
lint.select = [
  "B",  # flake8-bugbear - https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
  "C4", # flake8-comprehensions - https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
  "E",  # pycodestyle errors - https://docs.astral.sh/ruff/rules/#error-e
  "F",  # pyflakes rules - https://docs.astral.sh/ruff/rules/#pyflakes-f
  "W",  # pycodestyle warnings - https://docs.astral.sh/ruff/rules/#warning-w
  "I",  # isort - https://docs.astral.sh/ruff/rules/#isort-i
  "UP", # pyupgrade - https://docs.astral.sh/ruff/rules/#pyupgrade-up
]
